Separated _bonding into constituent components for greatly modularity and maintainability
